refactor: move flutter release model to separate file
